backend/requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
django-jazzmin==3.0.0  # Or latest version
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.0
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
channels==4.1.0  # Or latest stable version
# SMS and Email
twilio==8.13.0

# Deployment
gunicorn==21.2.0
dj-database-url==2.1.0
whitenoise==6.6.0
django-heroku==0.3.1
django-on-heroku==1.1.2

# AWS
boto3>=1.34.0
awscli>=1.32.0
botocore>=1.34.0
s3transfer>=0.10.0
django-s3-storage==0.15.0
aws-xray-sdk>=2.12.0
django-ses>=3.5.0

# Monitoring and logging
sentry-sdk==1.40.5
django-storages==1.14.2

# API Integrations
requests==2.31.0

# Calendar Integration
icalendar==6.3.1
